Revert "rename logger header"
This reverts commit cfe05c20721f92e4574f176cdc59f8f82e58c4cd.
