+++ stitcher/include/airmap/logger.h
